Fix #32 uninomials do not match
Now we check match results stems for exact match. That allows to weed
out false positives from bloom filters and gives us ability to trust
match results independently of gnames. Because of that we have to check
"stems" of uninmials as well, and becasue of that we have to add them to
the stem trie. The downside is that we completely unable to detect now
situations where space was lost or introduced by OCR. For that we need
to add another trie for non-stemmed canonicals, and I am not ready to do
that yet.